refactor(map): moved util functions to utils
Moved above 250 lines of code from `MapScreen.kt` to different `utils` files.

This is a first tentative at refactoring `MapScreen.kt` to make it more maintanable, a better work will be done in a future sprint.

For now, moving those functions out allows me to test them more easily.
